JNCTLASER -
Jinan Changtai CNC Technology Co., Ltd. is a professional laser cutting machine manufacturer  located in Jinan,Shandong province. which specializes in the manufacture and development of CNC machinery as well as the sales of domestic and foreign and technical service. it is engaged in the research, development, production and service of laser cutting machine , laser welding machine, laser marking machine, laser cleaning machine .
As a professional fiber laser cutting machine manufacturer,  we supply the CNC laser metal cutting equipment with high quality, high speed, high precision and high efficiency. It is suitable for all kinds of metal cutting which will be your good metal working partner. It is equipped with different powers (from 1000W to 30000W) for cutting metal sheets/plates and metal tubes/pipes with different thicknesses.